Added a generic data purge stored procedure to purge data
The stored procedure can be used on any table with a timestamp column.
It will delete data in the specified table with the specified timestamp column
having value older than the specified number of time unit from current timestamp.

Signature:

  purge_data(
    name_of_table text,
    name_of_timestamp_column text,
    number_of_time_unit_from_now integer,
    time_unit text)
    RETURNS void

Example Usage is to (manual or schedule) execute the stored procedure:

  SELECT purge_data('data_sample_raw', 'time_stamp', 6, 'MONTH');