Added the enable_pg_partition.py helper script for partition 'data_sample_raw' table
The enabled_pg_partition.py hardcoded the step to disable partition if exists
and enable partition on 'data_sample_raw' table using the 'time_stamp' column by
'DAY', and only keep the last 10 DAY of data.

These values can be changed in the script for different interval and number of
intervals to keep.

This python script requires SQLAlchemy python module to be installed and PostgreSQL
database driver 'psycopy2' for connecting to the database.

The script also expect an environment variable, 'PG_DB_URL', to contains the database
connection string in the format that SQLAlchemy expects.

    export PG_DB_URL=dialect+driver://username:password@host:port/database

http://docs.sqlalchemy.org/en/latest/core/engines.html#database-urls