Wrapped the create new partition statements with exception handler
Fix for bug CCS-741

Multiple aggregates may attempt to create the same partition and index
on the partition. A corner case is a race condition when two or more
aggregators are trying to insert tuples with the timestamp that required
the same partition to be created.

The 'IF NOT EXISTS' expression is sufficient to handle the concurrency
in 'CREATE TABLE' statement, but the 'CREATE INDEX' statement does not
support the 'IF NOT EXISTS' expression until PostgreSQL >= 9.5. The
solution here is to catch the exception and ignore it.

Again this is to fix a rare case of race condition, and the ELSE branch
of the IF statement only execute when a new partition needs to be created.